Adicionar metodo de apagar user ao front end UC120 #104 (G55-30)

+++ src/controllers/userController.ts
@@ -9,55 +9,55 @@ import IUserDTO from '../dto/IUserDTO';
 import { Result } from "../core/logic/Result";
 import ICreatingUserDTO from '../dto/ICreatingUserDTO';
 import { Console } from 'console';
+var authorizeEmail = require("../api/middlewares/validateToken")
 
 @Service()
 export default class UserController implements IUserController /* TODO: extends ../core/infra/BaseController */ {
   constructor(
-      @Inject(config.services.user.name) private userServiceInstance : IUserService
-  ) {}
+    @Inject(config.services.user.name) private userServiceInstance: IUserService
+  ) { }
 
-  public async criarUser(req: Request, res: Response, next: NextFunction){
+  public async criarUser(req: Request, res: Response, next: NextFunction) {
     try {
       const userOrError = await this.userServiceInstance.criarUser(req.body as ICreatingUserDTO) as Result<IUserDTO>;
-        
+
       if (userOrError.isFailure) {
         return res.status(403).send("Erro: " + userOrError.errorValue());
       }
 
       const userDTO = userOrError.getValue();
-      return res.json( userDTO ).status(201);
+      return res.json(userDTO).status(201);
     }
     catch (e) {
       return next(e);
     }
   };
 
-  public async criarUtente(req: Request, res: Response, next: NextFunction){
+  public async criarUtente(req: Request, res: Response, next: NextFunction) {
     try {
       const userOrError = await this.userServiceInstance.criarUtente(req.body as ICreatingUserDTO) as Result<IUserDTO>;
-        
+
       if (userOrError.isFailure) {
         return res.status(403).send("Erro: " + userOrError.errorValue());
       }
 
       const userDTO = userOrError.getValue();
-      return res.json( userDTO ).status(201);
+      return res.json(userDTO).status(201);
     }
     catch (e) {
       return next(e);
     }
   };
 
-  public async updateUser(req: Request, res: Response, next: NextFunction){
+  public async updateUser(req: Request, res: Response, next: NextFunction) {
     try {
       const userOrError = await this.userServiceInstance.updateUser(req.params.email, req.body as IUserDTO) as Result<IUserDTO>;
-        
       if (userOrError.isFailure) {
         return res.status(404).send("Erro: " + userOrError.errorValue());
       }
 
       const userDTO = userOrError.getValue();
-      return res.json( userDTO ).status(200);
+      return res.json(userDTO).status(200);
     }
     catch (e) {
       return next(e);
@@ -96,7 +96,11 @@ export default class UserController implements IUserController /* TODO: extends 
 
   public async deleteUser(req: Request, res: Response, next: NextFunction) {
     try {
-      const usersOrError = await this.userServiceInstance.deleteUser(req.params.email) as Result<IUserDTO>;
+      const stat = authorizeEmail(req.params.email);
+      let usersOrError;
+      if (stat.status(200)) {
+        usersOrError = await this.userServiceInstance.deleteUser(req.params.email) as Result<IUserDTO>;
+      }
 
       if (usersOrError.isFailure) {
         return res.status(404).send("Erro: " + usersOrError.errorValue());
@@ -111,7 +115,7 @@ export default class UserController implements IUserController /* TODO: extends 
 
   public async login(req: Request, res: Response, next: NextFunction) {
     try {
-      const usersOrError = await this.userServiceInstance.login(req.params.email,req.params.password) as Result<IUserDTO>;
+      const usersOrError = await this.userServiceInstance.login(req.params.email, req.params.password) as Result<IUserDTO>;
 
       if (usersOrError.isFailure) {
         return res.status(404).send("Erro: " + usersOrError.errorValue());
@@ -123,22 +127,22 @@ export default class UserController implements IUserController /* TODO: extends 
       return next(e);
     }
   };
-/*
-exports.getMe = async function(req, res: Response) {
+  /*
+  exports.getMe = async function(req, res: Response) {
+    
+      // NB: a arquitetura ONION não está a ser seguida aqui
   
-    // NB: a arquitetura ONION não está a ser seguida aqui
-
-    const userRepo = Container.get(config.repos.user.name) as IUserRepo
-
-    if( !req.token || req.token == undefined )
-        return res.json( new Error("Token inexistente ou inválido")).status(401);
-
-    const user = await userRepo.findById( req.token.id );
-    if (!user)
-        return res.json( new Error("Utilizador não registado")).status(401);
-
-    const userDTO = UserMap.toDTO( user ) as IUserDTO;
-    return res.json( userDTO ).status(200);
-}
-*/
+      const userRepo = Container.get(config.repos.user.name) as IUserRepo
+  
+      if( !req.token || req.token == undefined )
+          return res.json( new Error("Token inexistente ou inválido")).status(401);
+  
+      const user = await userRepo.findById( req.token.id );
+      if (!user)
+          return res.json( new Error("Utilizador não registado")).status(401);
+  
+      const userDTO = UserMap.toDTO( user ) as IUserDTO;
+      return res.json( userDTO ).status(200);
+  }
+  */
 }
